move function to file(BackwardWeightV4R4Helper.h)
