+++ mlir/include/mlir/Dialect/MIOpen/utility/common.hpp
@@ -1,0 +1,8 @@
+#ifndef COMMON_H
+#define COMMON_H
+struct ArgumentFields {
+  int gridwiseGemmArgumentPosition[3];
+  StringRef gemmTargetCharName[3];
+};
+
+#endif
